+++ HISTORY.md
@@ -1,10 +1,17 @@
 # Change log
 
-## 0.1.1 (08/27/2015)
+## Ang 0.1.0 (03/17/2017)
+
+* Changed: Renamed the project, reset the version number back to 0.1.0.
+* Feature: Add `impl PartialOrd/Ord` for Angle.
+* Feature: Add `impl approx::ApproxEq` from the [approx](http://crates.io/crates/approx) crate.
+* Changed: Use the `num-traits` crate instead of the num `crate`.
+
+## Angular-rust 0.1.1 (08/27/2015)
 
 * Feature: Add method `Angle::min_dist()` which computes the minimal
   unsigned distance between two normalized angles
 
-## 0.1.0 (08/26/2015)
+## Angular-rust 0.1.0 (08/26/2015)
 
 * Birth!